requirements.txt
Flask
spacy
scikit-learn
python-docx
PyPDF2
